+++ docs/conf.py
@@ -103,5 +103,5 @@ def setup(app) -> None:
     app.add_css_file("my_theme.css")
 
 
-# TODO: to be used with sphinx>=7.1
+# Only works with sphinx>=7.1
 maximum_signature_line_length = 10