+++ src/aac_metrics/__init__.py
@@ -6,20 +6,22 @@
 
 __author__ = "Etienne Labbé (Labbeti)"
 __author_email__ = "labbeti.pub@gmail.com"
+__docs__ = "Audio Captioning Metrics"
+__docs_url__ = "https://aac-metrics.readthedocs.io/en/stable/"
 __license__ = "MIT"
 __maintainer__ = "Etienne Labbé (Labbeti)"
 __name__ = "aac-metrics"
 __status__ = "Development"
-__version__ = "0.5.3"
+__version__ = "0.5.4"
 
 
 from .classes.base import AACMetric
 from .classes.bert_score_mrefs import BERTScoreMRefs
 from .classes.bleu import BLEU
 from .classes.cider_d import CIDErD
-from .classes.evaluate import Evaluate, DCASE2023Evaluate, _get_metric_factory_classes
-from .classes.fer import FER
+from .classes.evaluate import DCASE2023Evaluate, Evaluate, _get_metric_factory_classes
 from .classes.fense import FENSE
+from .classes.fer import FER
 from .classes.meteor import METEOR
 from .classes.rouge_l import ROUGEL
 from .classes.sbert_sim import SBERTSim
@@ -28,7 +30,7 @@ from .classes.spider import SPIDEr
 from .classes.spider_fl import SPIDErFL
 from .classes.spider_max import SPIDErMax
 from .classes.vocab import Vocab
-from .functional.evaluate import evaluate, dcase2023_evaluate
+from .functional.evaluate import dcase2023_evaluate, evaluate
 from .utils.globals import (
     get_default_cache_path,
     get_default_java_path,
@@ -37,7 +39,6 @@ from .utils.globals import (
     set_default_java_path,
     set_default_tmp_path,
 )
-
 
 __all__ = [
     "AACMetric",